pyproject.toml
[project]
name = "shhhitpost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.4",
    "httpx>=0.27.2",
    "pydantic-settings>=2.6.1",
    "pynacl>=1.5.0",
    "ruff>=0.7.2",
    "uvicorn>=0.32.0",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "T201"]
